Scope callout_stop drain case properly
A deadlock condition was introduced with commit 164ca60.
Essentially callout_stop was incorrectly always running the "drain"
code where it is allowed to block if curently executing.  Constrain
the behavior properly (via the CS_DRAIN flag), and cleanup one use
case.

Note: this likely restores the original race condition for which the
original set of changes were added to guard against.